Camera
Again, similar to the professor's camera function but spaced out a
little bit easier to read. Modified getRotatedCameraPosition() to take
an array rather than just an angle. I was attempting to pass height and
depth parameters as well as an angle parameter in the tick() function to
allow the camera to move in all directions.